requirements.txt
pandas
matplotlib
python-dotenv
newspaper3k
GoogleNews
transformers
torch
requests
tqdm